PhishNet-main/FlaskBack/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
future==1.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
MarkupSafe==2.1.3
numpy==2.4.0
python-dateutil==2.9.0.post0
pandas==2.3.3
scikit-learn==1.8.0
requests==2.31.0
requests-file==1.5.1
tldextract==5.3.0
whois==1.20240129.2
catboost==1.2.0
lightgbm>=4.1.0
optuna>=3.4.0
xgboost>=1.7.0
python-Levenshtein>=0.21.0
shap>=0.43.0
dnspython>=2.4.0
ipwhois>=1.3.0
selenium>=4.15.0
webdriver-manager>=4.0.0
scikit-image>=0.22.0
opencv-python>=4.8.0
Pillow>=10.0.0
